--- src/fallbackError.ts
@@ -1,94 +1,0 @@
-/**
- * Copyright 2020 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import * as protobuf from 'protobufjs';
-import {Status} from './status';
-
-// A copy of gRPC ServiceError but with no metadata
-export type FallbackServiceError = FallbackStatusObject & Error;
-interface FallbackStatusObject {
-  code: Status;
-  message: string;
-  details: Array<{}>;
-}
-
-interface ProtobufAny {
-  type_url: string;
-  value: Uint8Array;
-}
-
-interface RpcStatus {
-  code: number;
-  message: string;
-  details: ProtobufAny[];
-}
-
-export class FallbackErrorDecoder {
-  root: protobuf.Root;
-  anyType: protobuf.Type;
-  statusType: protobuf.Type;
-
-  constructor() {
-    // eslint-disable-next-line @typescript-eslint/no-var-requires
-    const errorProtoJson = require('../../protos/status.json');
-    this.root = protobuf.Root.fromJSON(errorProtoJson);
-    this.anyType = this.root.lookupType('google.protobuf.Any');
-    this.statusType = this.root.lookupType('google.rpc.Status');
-  }
-
-  decodeProtobufAny(anyValue: ProtobufAny): protobuf.Message<{}> {
-    const match = anyValue.type_url.match(/^type.googleapis.com\/(.*)/);
-    if (!match) {
-      throw new Error(
-        `Unknown type encoded in google.protobuf.any: ${anyValue.type_url}`
-      );
-    }
-    const typeName = match[1];
-    const type = this.root.lookupType(typeName);
-    if (!type) {
-      throw new Error(`Cannot lookup type ${typeName}`);
-    }
-    return type.decode(anyValue.value);
-  }
-
-  // Decodes gRPC-fallback error which is an instance of google.rpc.Status.
-  decodeRpcStatus(buffer: Buffer | ArrayBuffer): FallbackStatusObject {
-    const uint8array = new Uint8Array(buffer);
-    const status = this.statusType.decode(uint8array) as unknown as RpcStatus;
-
-    // google.rpc.Status contains an array of google.protobuf.Any
-    // which need a special treatment
-    const result = {
-      code: status.code,
-      message: status.message,
-      details: status.details.map(detail => this.decodeProtobufAny(detail)),
-    };
-    return result;
-  }
-
-  // Construct an Error from a StatusObject.
-  // Adapted from https://github.com/grpc/grpc-node/blob/master/packages/grpc-js/src/call.ts#L79
-  callErrorFromStatus(status: FallbackStatusObject): FallbackServiceError {
-    status.message = `${status.code} ${Status[status.code]}: ${status.message}`;
-    return Object.assign(new Error(status.message), status);
-  }
-
-  // Decodes gRPC-fallback error which is an instance of google.rpc.Status,
-  // and puts it into the object similar to gRPC ServiceError object.
-  decodeErrorFromBuffer(buffer: Buffer | ArrayBuffer): Error {
-    return this.callErrorFromStatus(this.decodeRpcStatus(buffer));
-  }
-}
